Rename autoconfigure package
From `com.kakawait.spring.boot.security.cas` to `com.kakawait.spring.boot.security.cas.autoconfigure`

fixes #33
